+++ Gw2Sharp/WebApi/V2/Clients/Pvp/IPvpSeasonsLeaderboardsRegionIdClient.cs
@@ -6,7 +6,6 @@ namespace Gw2Sharp.WebApi.V2.Clients
     /// A client of the Guild Wars 2 API v2 PvP season leaderboard region endpoint.
     /// </summary>
     public interface IPvpSeasonsLeaderboardsRegionIdClient :
-        IBlobClient<IApiV2ObjectList<PvpSeasonLeaderboardEntry>>
-    {
-    }
+        IBlobClient<IApiV2ObjectList<PvpSeasonLeaderboardEntry>>,
+        IPaginatedClient<PvpSeasonLeaderboardEntry> { }
 }