+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Gw2Sharp History
 
 ## 0.11.0
+### Endpoints
+- Update `Gw2Sharp.WebApi.V2.Pvp.Seasons[id].Leaderboards[board][region]` to support pagination ([#55](https://github.com/Archomeda/Gw2Sharp/issues/55))
+
 ### Fixes
 - Fix possible memory leak when using the archive memory cache in combination with binary data (a `MemoryStream` was not being disposed)
 - Fix `KeyNotFoundException` when performing a many request with at least one (but not all) invalid id ([#65](https://github.com/Archomeda/Gw2Sharp/issues/65), [#66](https://github.com/Archomeda/Gw2Sharp/pull/66))
+- Fix `LegendType` to start with 14 instead of 13 ([#63](https://github.com/Archomeda/Gw2Sharp/issues/63))
+- Fail early when an access token is incorrectly formatted by throwing an `ArgumentException` during constructing of the `Connection` object, and when setting it through the `Connection.AccessToken` property ([#62](https://github.com/Archomeda/Gw2Sharp/issues/62))
 
 ### Refactoring
 - **Breaking:** `Gw2Sharp.WebApi.V2.Models.Profession.SkillsByPalette` is now using a `IReadOnlyDictionary<int, int>` instead of `IDictionary<int, int>`
@@ -13,6 +18,7 @@
   - The `Gw2Sharp.Json.Converters.CompactMapConverter` that's used for endpoints that return objects with skill palettes has been changed to deserialize to `IReadOnlyDictionary<K, V>` instead of `IDictionary<K, V>`
 - A bunch of internal code changes to conform ReSharper and Microsoft.CodeAnalysis.FxCopAnalyzers
 - All methods that compare strings, parse strings or create strings from other types have been updated to make sure that they ignore the system locale
+- Mount type 10 in Mumble Link doesn't redirect to `MountType.None` anymore ([#64](https://github.com/Archomeda/Gw2Sharp/issues/64))
 
 ---